# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/buy_method/api/v1/": {
            "get": {
                "description": "Get all buymethods",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "buymethod"
                ],
                "summary": "List of buymethods",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by age",
                        "name": "age",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/buy_method.BuyMethodResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new buymethod with the provided information",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "buymethod"
                ],
                "summary": "Create buymethod",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BuyMethod name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "BuyMethod age",
                        "name": "age",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "BuyMethod image",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/buy_method.BuyMethodResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/basics.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/basics.APIError"
                        }
                    }
                }
            }
        },
        "/buy_method/api/v1/{id}": {
            "get": {
                "description": "Retrieve details of a buymethod by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "buymethod"
                ],
                "summary": "Get buymethod details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BuyMethod ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/buy_method.BuyMethodResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid UUID format",
                        "schema": {
                            "$ref": "#/definitions/basics.APIError"
                        }
                    },
                    "404": {
                        "description": "BuyMethod not found",
                        "schema": {
                            "$ref": "#/definitions/basics.APIError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update buymethod details by ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "buymethod"
                ],
                "summary": "Update buymethod",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BuyMethod ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "BuyMethod name",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "BuyMethod age",
                        "name": "age",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "BuyMethod image",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/buy_method.BuyMethodResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/basics.APIError"
                        }
                    },
                    "404": {
                        "description": "BuyMethod not found",
                        "schema": {
                            "$ref": "#/definitions/basics.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/basics.APIError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a buymethod by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "buymethod"
                ],
                "summary": "Delete buymethod",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BuyMethod ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid UUID format",
                        "schema": {
                            "$ref": "#/definitions/basics.APIError"
                        }
                    },
                    "404": {
                        "description": "BuyMethod not found",
                        "schema": {
                            "$ref": "#/definitions/basics.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/basics.APIError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update specific fields of a city by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "buymethod"
                ],
                "summary": "Update city partially",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BuyMethod ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Partial BuyMethod information",
                        "name": "city",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/buy_method.CreateBuyMethodRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/buy_method.BuyMethodResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/basics.APIError"
                        }
                    },
                    "404": {
                        "description": "buymethod not found",
                        "schema": {
                            "$ref": "#/definitions/basics.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/basics.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "basics.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "buy_method.BuyMethodResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "persianName": {
                    "type": "string"
                }
            }
        },
        "buy_method.CreateBuyMethodRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "persianName": {
                    "type": "string"
                }
            }
        },
...
    }
}